chore: move FirebaseFileUploader relative to reference
